refactor to make code simpler and more reusable for home delivery later
